Move /images to top level feature.
